+++ sourceCode/scripting/userParameters.h
@@ -22,11 +22,11 @@ public:
     void clearInitialParameters();
 
     // For following 2 functions, the paramName must be present for the function to return true:
-    bool setParameterValue(const char* paramName,const char* paramValue,int paramValueLength); // paramValue can contain embedded 0's since 2011/01/23
+    bool setParameterValue(const char* paramName,const char* paramValue,size_t paramValueLength); // paramValue can contain embedded 0's since 2011/01/23
     bool getParameterValue(const char* paramName,std::string& paramValue);
 
     // For following 3 functions, the paramName does not need to be present:
-    void addParameterValue(const char* paramName,const char* unitText,const char* paramValue,int paramValueLength); // paramValue can contain embedded 0's since 2011/01/23
+    void addParameterValue(const char* paramName,const char* unitText,const char* paramValue,size_t paramValueLength); // paramValue can contain embedded 0's since 2011/01/23
     bool removeParameterValue(const char* paramName);
     bool removeParameterValue(int index);
     int getParameterIndex(const char* paramName);
